Inicio da Migração para .net 6
