+++ src/API/Controllers/WeatherForecastController.cs.bak
@@ -14,8 +14,8 @@ namespace API.Controllers
     {
         private static readonly string[] Summaries = new[]
         {
-        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-    };
+            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+        };
 
         private readonly ILogger<WeatherForecastController> _logger;
 
@@ -28,7 +28,8 @@ namespace API.Controllers
         private IModel model;
 
 
-        [HttpGet(Name = "GetWeatherForecast")]
+        [HttpGet]
+        [Route("weatherForecast")]
         public IEnumerable<WeatherForecast> Get()
         {
             var rng = new Random();
@@ -42,15 +43,16 @@ namespace API.Controllers
             .ToArray();
         }
 
-        [HttpPost(Name ="weatherForecast")]
+        [HttpPost]
+        [Route("weatherForecast")]
         public IEnumerable<WeatherForecast> Post()
         {
             var rng = new Random();
             return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
                 Date = DateTime.Now.AddDays(index),
-                TemperatureC = Random.Shared.Next(-20, 55),
-                Summary = Summaries[Random.Shared.Next(Summaries.Length)]
+                TemperatureC = rng.Next(-20, 55),
+                Summary = Summaries[rng.Next(Summaries.Length)]
             })
             .ToArray();
         }
