+++ src/API/WeatherForecast.cs.bak
@@ -12,6 +12,6 @@ namespace API
 
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-        public string? Summary { get; set; }
+        public string Summary { get; set; }
     }
 }
