+++ src/KeyCloak/docker-compose.yml
@@ -119,6 +119,7 @@ services:
     image: quay.io/keycloak/keycloak:14.0.0
     container_name: keycloak
     environment:
+      - PROXY_ADDRESS_FORWARDING=true
       - DB_VENDOR=postgres
       - DB_USER=postgres
       - DB_PASSWORD=postgres
