+++ src/KeyCloak/nginxconfig/keycloak.conf
@@ -14,7 +14,9 @@ server {
     ssl_session_tickets         off;
     ssl_stapling                on;
     ssl_stapling_verify         on;
-    add_header Strict-Transport-Security max-age=31536000;
+    add_header Strict-Transport-Security "max-age=31536000; includeSubDomains";
+    add_header X-XSS-Protection "1; mode=block";
+    add_header Content-Security-Policy "frame-src 'self'; frame-ancestors 'self'; object-src 'none';";
     server_name  keycloak.localhost;
 
         #Begin Extensions        
@@ -27,10 +29,12 @@ server {
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Host $server_name;
+            proxy_set_header Cookie $http_cookie;
             proxy_cache off;
             proxy_buffers         8 16k;
             proxy_buffer_size       16k;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             gzip on;
             proxy_redirect http:// https://; 
         }
@@ -41,7 +45,9 @@ server {
     listen       9090;
     listen  [::]:9090;
     server_name   keycloak.localhost ;
-    add_header Strict-Transport-Security max-age=31536000;
+    add_header Strict-Transport-Security "max-age=31536000; includeSubDomains";
+    add_header X-XSS-Protection "1; mode=block";
+    add_header Content-Security-Policy "frame-src 'self'; frame-ancestors 'self'; object-src 'none';";
 
         #Begin Extensions        
     add_header Feedback identity-server_keycloak;
@@ -52,10 +58,12 @@ server {
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-Proto $scheme;
+            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            proxy_set_header X-Forwarded-Host $server_name;
+            proxy_set_header Cookie $http_cookie;
             proxy_cache off;
             proxy_buffers         8 16k;
             proxy_buffer_size       16k;
-            proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             gzip on;
         }
