+++ src/MVC/Dockerfile
@@ -1,6 +1,6 @@
 #See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
 
-FROM mcr.microsoft.com/dotnet/core/aspnet:3.1-buster-slim AS base
+FROM  mcr.microsoft.com/dotnet/aspnet:6.0 AS base
 ENV TZ=America/Sao_Paulo
 RUN cat /usr/share/zoneinfo/$TZ > /etc/localtime \
 		&& cat /usr/share/zoneinfo/$TZ > /etc/timezone
@@ -8,7 +8,7 @@ WORKDIR /app
 EXPOSE 80
 EXPOSE 443
 
-FROM mcr.microsoft.com/dotnet/core/sdk:3.1-buster AS build
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
 WORKDIR /src
 COPY ./Cadastro.MVC.csproj MVC/
 RUN dotnet restore ./MVC/Cadastro.MVC.csproj
